Broke tests into smaller files
Moved test describe groups into smaller file chunks.
Apply toxy only to the end to end tests.
